feat: add response struct for json response
