+++ src/models/manga.rs
@@ -86,71 +86,3 @@ impl<'a> NewManga<'a> {
             .get_result(connection)
     }
 }
-
-// TODO: put resp into separate file or into more proper file
-#[derive(Serialize)]
-pub struct Resp {
-    pub manga:    Manga,
-    pub staff:    Vec<crate::models::staff::Staff>,
-    pub relation: Vec<crate::models::relation::Relation>,
-}
-
-impl Manga {
-    pub fn full(
-        mid: String,
-        connection: &PgConnection,
-    ) -> Result<Resp, diesel::result::Error> {
-        use crate::diesel::ExpressionMethods;
-        use crate::diesel::QueryDsl;
-        use crate::diesel::RunQueryDsl;
-        use crate::models::{
-            media::Media,
-            relation::Relation,
-            series::Series,
-            staff::Staff,
-        };
-        use crate::schema::{
-            media,
-            relations,
-            series,
-            staffs,
-        };
-
-        let uid = uuid::Uuid::parse_str(&mid).unwrap();
-        let manga: Manga = mangas::table
-            .filter(mangas::id.eq(uid))
-            .get_result(connection)?;
-        let series: Result<Vec<Series>, diesel::result::Error> = series::table
-            .filter(series::manga_id.eq(uid))
-            .get_results(connection);
-        let staff_vec: Vec<Staff> = series
-            .into_iter()
-            .flatten()
-            .flat_map(|k| {
-                staffs::table
-                    .filter(staffs::id.eq(k.staff_id))
-                    .get_results(connection)
-                    .unwrap()
-            })
-            .collect::<Vec<Staff>>();
-        let relations: Result<Vec<Media>, diesel::result::Error> = media::table
-            .filter(media::manga_id.eq(uid))
-            .get_results(connection);
-        let relation_vec: Vec<Relation> = relations
-            .into_iter()
-            .flatten()
-            .flat_map(|r| {
-                relations::table
-                    .filter(relations::id.eq(r.relation_id))
-                    .get_results(connection)
-                    .unwrap()
-            })
-            .collect::<Vec<Relation>>();
-
-        Ok(Resp {
-            manga:    manga,
-            staff:    staff_vec,
-            relation: relation_vec,
-        })
-    }
-}
