+++ src/models/mod.rs
@@ -2,5 +2,6 @@ pub mod json_manga;
 pub mod manga;
 pub mod media;
 pub mod relation;
+pub mod response;
 pub mod series;
 pub mod staff;
